Explicitly set content type on xhr request
As a user I had to introduce a proxy between the plugin and
Anki actual program to work around issues with Snap packaging
(Having Obsidian installed as a snap package).

That proxy used to set default headers on the request and
caused it to be processed incorrectly. This line prevented
that for the content type header.

It's important because when AnkiConnect does not recognize
the content type it returns something like "AnkiConnect v.6"
which cause the connection to fail.